there is only insertion to fix, and it is on data file insertion, renan pls help

+++ projeto_2/src/FileManager/FileWalker/file_walker.h
@@ -24,7 +24,7 @@ int fw_get_reg_number(FileWalker fw);
 void add_removed_list(FileWalker fw, Register reg);
 void debug_removed_list(FileWalker fw);
 
-int fw_insert_into(FileWalker fw, Register reg);
+long int fw_insert_into(FileWalker fw, Register reg);
 
 int fw_print_with_offset(FileWalker fw, Filter filter, long int offset);
